==> src/07_figures_charts_Discussion.Rmd <==
---
title: "07_figures_charts"
author: "Kolja Nenoff"
date: "2023-07-08"
output: html_document
---
```{r setup, include=FALSE}
rm(list = ls())
library(rstudioapi)
library(magrittr)
library(tidyverse)
library(data.table)
library(parallel)
library(dimRed)
library(devtools)
library(igraph)
library(energy)
library(ggrepel)
library(ggalt)
library(ggraph)
library(trend)
library(riverplot)
library(RANN)
library(RJSONIO)
library(RSpectra)
library(pcaMethods)
require(broom)
require(GGally)
require(ggpubr)
require(xgboost)
require(caret)
require(pdp)
require(iml)
require(data.table)
require(tidyverse)
require(cowplot)
require(readxl)
```
```{r}
wdi_explained<-fread("../data/WDI_data_230609/WDI_CSV/WDISeries.csv")
#load("../data/07_230504_OWID_dataset")
dimred2021_knn250_2<-fread("../output/02_230612_Dimred_ExcessMort_annual.csv")
col_pal <- c("#1F968B",  "#7570b3", "#e7298a", "#3CBB75","yellow3", "orange","slategrey")
```
# functions
```{r}
source("00_toolbox.R")
source("../libraries/01a_lib_social_indicators.R")

```

# data
```{r,Excess Mortility WHO Set}

excessCounrty_year<-read_excel("../../01_data_prep/data/set_mortality_230221/230224_nature_WHO_excess_death/2022-03-25_covid-19_gem/WHO_COVID_Excess_Deaths_EstimatesByCountry.xlsx",sheet = 2)
excessCounrty_month<-read_excel("../../01_data_prep/data/set_mortality_230221/230224_nature_WHO_excess_death/2022-03-25_covid-19_gem/WHO_COVID_Excess_Deaths_EstimatesByCountry.xlsx",sheet = 2)
# excessCounrty_covariates<-read_excel("../data/230224_nature_WHO_excess_death/2022-03-25_covid-19_gem/WHO_COVID_Excess_Deaths_EstimatesByCountry.xlsx",sheet = 3)
# excessRegion_year<-read_excel("../data/230224_nature_WHO_excess_death/2022-03-25_covid-19_gem/WHO_COVID_Excess_Deaths_EstimatesByRegion.xlsx",sheet = 2)
# excessRegion_income<-read_excel("../data/230224_nature_WHO_excess_death/2022-03-25_covid-19_gem/WHO_COVID_Excess_Deaths_EstimatesByIncome.xlsx",sheet = 2)
# excessRegion_income %>% group_by(income) %>% 
#   summarise(p_value_mort=sum(excess.mean)/sum(expected.mean)*100,excess_comul=sum(excess.mean))
excessCounrty_year_pvalue<-excessCounrty_year %>% group_by(country,iso3,year) %>% 
  summarise(p_value_mort=sum(excess.mean)/sum(expected.mean)*100,excess_comul=sum(excess.mean))
excessCounrty_year %>% filter(year==2020)-> excessCounrty_year_2020
p_value_global_2020=sum(excessCounrty_year_2020$excess.mean,na.rm = T)/sum(excessCounrty_year_2020$expected.mean,na.rm = T)*100
excessCounrty_year %>% filter(year==2021)-> excessCounrty_year_2020
p_value_global_2021=sum(excessCounrty_year_2020$excess.mean,na.rm = T)/sum(excessCounrty_year_2020$expected.mean,na.rm = T)*100
excessCounrty_monthly_pvalue<-excessCounrty_year  %>% 
  mutate(p_value_mort=(excess.mean)/(expected.mean)*100)
# 
# excessmonthly <- excessCounrty_monthly_pvalue %>% arrange(country,year,month) %>% rename(`Country Code`=iso3) %>% 
#   mutate(date = as.Date(paste(year, month, "01", sep = "-"), format = "%Y-%m-%d"),identifier=paste0(`Country Code`,"_",year(date),"_",month(date)))%>% filter(date>"2020-02-01") %>% merge( . , thisisOWID,by.x="identifier",by.y="monthyear")
```


# Method 1 Compare ISO and PCA
```{r}
## set up path
DATA_DIR <- "../data/"
FIGURE_DIR <- "../figures/01_figures/"
WDI_DIR <- paste0(DATA_DIR, "WDI_data_230609/WDI_CSV")
RELATIVE_INDICATORS_FILE_NAME <- "01_data/relative_indicators_2023_06.R" # nolint
source("../libraries/01a_lib_social_indicators.R")


residual_sample_variances <- readRDS("../data/residual_sample_variances.RDS")
cons_iso_pca<-readRDS("../data/cons_iso_pca.RDS")
occurrence_table       <- cons_iso_pca$occurrence_table
wdi_data_cons          <- cons_iso_pca$wdi_data_cons
wdi_data_cons_pca      <- cons_iso_pca$wdi_data_cons_pca
wdi_data_cons_iso      <- cons_iso_pca$wdi_data_cons_iso
wdi_data_qual_cons_pca <- cons_iso_pca$wdi_data_qual_cons_pca
wdi_data_qual_cons_iso <- cons_iso_pca$wdi_data_qual_cons_iso
knns                   <- cons_iso_pca$knns
knns
cons_iso_pca$wdi_data_cons_iso %>%  head
wdi_data_qual_cons_iso
```


```{r}
pca_res <- residual_sample_variances$pca
iso_res <- residual_sample_variances$iso
median_res_var_pca <- residual_sample_variances$median_pca
median_res_var_iso <- residual_sample_variances$median_iso
hdi_res_var <- residual_sample_variances$hdi

plot_if_no_file(
  get_plot_file("01_230708_pca_iso_res_var.pdf"),
  plot_residual_variances,
  residual_sample_variances = residual_sample_variances,
  wdi_data_qual_cons_iso = wdi_data_qual_cons_iso,
  wdi_data_qual_cons_pca = wdi_data_qual_cons_pca,
  .overwrite = TRUE
)
```
```{r}
residual_sample_variances = residual_sample_variances
wdi_data_qual_cons_pca = wdi_data_qual_cons_pca
wdi_data_qual_cons_iso = wdi_data_qual_cons_iso
  
old_par <- par()
par(mar = c(3.5, 3.75, 0, 0) + 0.1, las = 1, mgp = c(2.5, 0.25, 0))
on.exit(par(old_par))

max_x <- 14

pca_res <- residual_sample_variances$pca
iso_res <- residual_sample_variances$iso
median_res_var_pca <- residual_sample_variances$median_pca
median_res_var_iso <- residual_sample_variances$median_iso
hdi_res_var <- residual_sample_variances$hdi

max_res_var_1 <- 1
png("../figures/06_plots/myplot.png",width = 720,height = 720,units = "px")
plot(pca_res[[1]],
     ylim = c(0, max_res_var_1),
     ## xlim = c(1, max_dim),
     xlim = c(1, max_x),
     ylab = "residual variance",
     xlab = "number of dimensions",
     cex.lab = 2,
     las = 1,
     xaxt = "n",
     yaxt = "n",
     bty = "n",
     type = "n")
points(wdi_data_qual_cons_pca,      type = "b",
       lwd = 4, col = lighten(col_pal[7], 0.8))
points(wdi_data_qual_cons_iso[[2]], type = "b",
       lwd = 4, col = lighten(col_pal[6], 0.8))
text(1, wdi_data_qual_cons_pca[1],      "PCA",
     pos = 4, offset = 0.5, xpd = TRUE, cex = 2)
text(1, wdi_data_qual_cons_iso[[3]][1], "Ensemble Isomap",
     pos = 4, offset = 0.5, xpd = TRUE, cex = 2)
text(max_x + 0.75,         0.1, "10% Residual Variance", cex = 1.8,
     adj = c(1.1, -0.5), xpd = TRUE)
abline(h = 0.1, col = "gray50", lwd = 2)

par(las = 1, mgp = c(1.5, 0.75, 0))#, xpd = TRUE)
axis(side = 1, at = 1:max_x, labels = ifelse(1:14 %% 2 == 1, 1:14, ""),
     col = "gray50", col.axis = "gray50",
     cex.axis = 1.5, lwd = 1, tcl = -0.25, lwd.ticks = 2)
axis(side = 2, at = seq(0, 1, length.out = 6),
     col = "gray50", col.axis = "gray50",
     cex.axis = 1.5, lwd = 1, tcl = -0.25, lwd.ticks = 2)
dev.off
```
# Method 2 Visualisation of  Nonlinearity
```{r}

a<-ggplot(dimred2021_knn250_2) +geom_point(aes(x=iso1,y=SH.STA.MMRT),color=col_pal[1])+ theme_bw() +ylab("Maternal mortality ratio \n(per 100 k live births)")+ xlab("Dimension 1")
b<-ggplot(dimred2021_knn250_2) +geom_point(aes(x=iso1,y=NY.GDP.PCAP.CD),color=col_pal[2])+ theme_bw() +ylab(wdi_explained[wdi_explained$`Series Code`%in% "NY.GDP.PCAP.CD","Indicator Name"])+ xlab("Dimension 1")
library(gridExtra)
grid.arrange(a,b) %>%  ggsave(file="../figures/06_plots/b_methods_nonlinear_pattern.png",device = "png")
```

# Results 1 Shapley values and importance of nonlinear prediction of Xgboost 



```{r}
Xgboostinput<- fread("../results/06_Dimred_Exmort_monthly.csv")
Label<-"p_value_mort"
features<-c("iso1","iso2","iso3","iso4","iso5","iso6","iso7")


custom_control <- trainControl(
  method = "cv",
  number = 30,  # Increase the number of folds
  search = "random",
  allowParallel = TRUE
  
)

param_grid <- expand.grid(
  nrounds = c(50,100),
  eta = c(0.05),
  max_depth = c(2,3, 4),
  colsample_bytree = c( 1),
  subsample = c(0.9),
  gamma = c(0),
  min_child_weight = c(0)
)

# Set the proportion for the test set (e.g., 0.2 for 20% test data)

test_proportion <- 0.8
Xgboostinput_filtered<-Xgboostinput %>%  select(features,Label) %>% drop_na()
index <- caret::createDataPartition(Xgboostinput_filtered$iso1, p = test_proportion, list = FALSE)
# Split the data into training and test sets
train_data <- Xgboostinput_filtered[index, ]
test_data <- Xgboostinput_filtered[-index, ]


# Train the XGBoost model using the function
trainedModel<- train_xgboost_model(train_data, Label, features, custom_control, param_grid)
# Assuming you have the 'finalmodel' and 'test_data' from the previous steps
predictedValue<-postResample(
      predict(trainedModel, 
          test_data %>% dplyr::select(features)),  
        test_data %>% pull(Label)
      )
predictedValue

```


```{r}

plota<-xgboost::xgb.ggplot.shap.summary(data = as.matrix(train_data %>% 
  dplyr::select(features)),
  model = trainedModel$finalModel) +scale_color_viridis("Feature Value")+ theme_bw() + xlab("Features")+ylab("Shapley value")+theme(legend.position = "left")

subsetimportance<-varImp(trainedModel)$importance %>%
         mutate(var = rownames(varImp(trainedModel)$importance))%>%
         dplyr::rename(importance = Overall) 


plotb<-ggplot(subsetimportance,aes(x = importance,
    y = reorder(var, importance)),)+
  geom_col(fill="slategrey")+
  ylab("")+theme_classic()+theme(axis.text.y=element_blank(),)

plot_grid(plota, plotb, labels = "A",rel_widths =   c(4/5, 1/5))
ggsave("../figures/06_plots/ResultsA_nonlinearexmort.png",device = "png")
```
```{r}
importance <- xgb.importance(model = trainedModel$finalModel, data = train_data)

importance_dt <- data.table::as.data.table(importance)

# Plot the feature importances in a heatmap
xgb.plot.importance(importance_matrix = importance_dt, top_n = 10)
```

# Results 2 WDI or Dimred How predicts better bootstrap


### impute WDI
```{r}

shrinkWDI<-dimred2021_knn250_2[,colnames(dimred2021_knn250_2)%in% c("Country Code","year",wdi_explained$`Series Code`),with=F] %>% filter(year<2020)

aTable<-shrinkWDI[,"Country Code"] %>%  unique
for(i in 3:ncol(shrinkWDI)){
  look<-shrinkWDI[,colnames(shrinkWDI)[c(1,2,i)],with=F] 
  colnames(look)[3]<-"value"
  best<-look[,.(value=last(value),year=last(year)),by=`Country Code`] %>%  filter(!is.na(value))
  okaisch<-look[!(look$`Country Code`%in% best$`Country Code`)&year%in% c(2000:2020),.(value=mean(value,na.rm=T),sd=var(value,na.rm=T)),by=`Country Code`]
  
  thisone<-rbind(best[,c("Country Code","value")],okaisch[,c("Country Code","value")])
  colnames(thisone)[2]<-colnames(shrinkWDI)[c(i)]
  aTable<-left_join(aTable,thisone, by = "Country Code")
}

bTable<-aTable[,-1] %>% sapply( . ,function(x)(as.numeric(x))) %>% data.frame() 

cTable<-sapply(bTable, function(x)x<-if_else(is.nan(x),999,x)) %>% data.frame()

cTable[cTable==999]<-NA
df_filled <- pca(object =  cTable, method = "ppca")
table2019_wdiImputed<-left_join(cbind(aTable[,1],data.frame(df_filled@completeObs)),dimred2021_knn250_2[dimred2021_knn250_2$year%in%2020,c("Country Code","pscore_mean","year")],by = "Country Code")

```


```{r}

param_grid <- expand.grid(
  nrounds = c(50),
  max_depth = c(1,2,3),           # Maximum depth of a tree
  eta = c(0.05),
  gamma = c(0, 1, 5),               # Minimum loss reduction required to make a further partition on a leaf node (regularization term)
  colsample_bytree = c(1),
  min_child_weight = c(0),
  subsample = c(0.95)
)

custom_control <- trainControl(
  method = "cv",
  number = 30,  # Increase the number of folds
  search = "random",
  allowParallel = TRUE)
  
param_list <- list(objective = "reg:squarederror",  # For regression
                   eta = 0.05,
                   max_depth = c(1,2,3),
                   gamma = c(0, 1, 5),    
                   subsample = 0.95
                   )


```

```{r}




Xgboostinput_first<-merge(table2019_wdiImputed,excessCounrty_monthly_pvalue %>% select(p_value_mort,iso3),by.x = "Country Code",by.y="iso3")


data_wdi<-Xgboostinput_first
trials <- 1
cl <- makeCluster(20)

require(doParallel)
data_wdi<-table2019_wdiImputed %>% filter(!is.na(pscore_mean))
trials <- 1000
cl <- makeCluster(10)

registerDoParallel(cl)
# Evaluate the performance of the model using RMSE


system.time(result_bootstrap <- foreach(icount(trials),
                  .combine = bind_rows) %dopar% {
                    require(xgboost)
                    require(tidyverse)
                    require(data.table)
    train_index <- sample(1:nrow(data_wdi), 0.8 * nrow(data_wdi))
    features<-colnames(data_wdi)[!grepl("Country|year|p_value|pscore_mean|[1-9]{2}",colnames(data_wdi))]

    Label <- "p_value_mort"
    trainDAT <- data_wdi[train_index, ]
    testDAT <- data_wdi[-train_index, ]
    test_proportion <- 0.8
    

    
    # Train the XGBoost model using the function
    trainedModel<- train_xgboost_model(trainDAT, Label, features, custom_control, param_grid)
    # Assuming you have the 'finalmodel' and 'test_data' from the previous steps
    predictedValue<-postResample(
      predict(trainedModel, 
          testDAT %>% dplyr::select(features)),  
        testDAT %>% pull(Label)
      )
    predictedValue
    
    # xgb_caret <- caret::train(x = trainDAT %>% dplyr::select(all_of(features)),
    #                y = trainDAT %>% pull(label),
    #                method = "xgbTree",
    #                verbosity = 0,
    #                trControl=custom_control,
    #   tuneGrid = param_grid
    #                #,tuneLength = 2
    # 
    #                 )
    # 
    # finalmodel <- caret::train(x = trainDAT %>% dplyr::select(all_of(features)),
    #                y = trainDAT %>% pull(label),
    #                method = "xgbTree",
    #                verbosity = 0,
    #                ## CAST CV LLO (Hanna Meyer)
    #                trControl=trainControl(method="cv",
    #                                       number = 30),
    # 
    #                tuneGrid = xgb_caret$bestTune)
    # 
    # importance_matrix<-varImp(finalmodel)$importance %>%
    #      mutate(var = rownames(varImp(xgb_caret)$importance))%>%
    #      dplyr::rename(importance = Overall) 
   
    

    
}
)


result_bootstrap_dimred_2019toPScoreMean<-result_bootstrap
result_bootstrap
fwrite(result_bootstrap_dimred_2019toPScoreMean,file = "../results/07_wdi_19_to_pscoreMean_bootstrap_1000.csv")
# load("../data/06_wdi_20_bootstrap_1000")


```
### Dimred



```{r}

trials <- 1000
cl <- makeCluster(30)
registerDoParallel(cl)
IsoYear<-"2019"
label<-"pscore_mean"
features<-c("iso1","iso2","iso3","iso4","iso5","iso6","iso7")

dimredData<-dimred2021_knn250_2[] %>% filter(year%in% 2019) %>% select(c(label,features)) %>%  filter(!is.na(iso1),!is.na(pscore_mean))
features<-c("iso1","iso2","iso3","iso4","iso5","iso6","iso7")

label<-"pscore_mean"

system.time(result_bootstrap_dimred <- foreach(icount(trials),
                  .combine = bind_rows) %dopar% {
                    require(xgboost)
                    require(tidyverse)
                    require(data.table)
                    library(shapper)
    
                    
    train_index <- sample(1:nrow(dimredData), 0.8 * nrow(dimredData))
    features<-c("iso1","iso2","iso3","iso4","iso5","iso6","iso7")

    label<-"pscore_mean"
    
    trainDAT <- dimredData[train_index, ]
    testDAT <- dimredData[-train_index, ]
    
    xgb_caret_dimred <- caret::train(x = trainDAT %>% dplyr::select(all_of(features)),
                   y = trainDAT %>% pull(label),
                   method = "xgbTree",
                   verbosity = 0,
                   trControl=custom_control,tuneGrid = param_grid
                   ,tuneLength = 1
   
                    )
    
    
    finalmodel_dimred <- caret::train(x = trainDAT %>% dplyr::select(all_of(features)),
                   y = trainDAT %>% pull(label),
                   method = "xgbTree",
                   verbosity = 0,
                   ## CAST CV LLO (Hanna Meyer)
                   trControl=trainControl(method="cv",
                                          number = 30),
                   # early_stopping_rounds = 10,
                   tuneGrid = xgb_caret_dimred$bestTune)
    
    importance_matrix<-varImp(finalmodel_dimred)$importance %>%
         mutate(var = rownames(varImp(xgb_caret_dimred)$importance))%>%
         dplyr::rename(importance = Overall) 
   

    thisone<-length(list.files("../test/dimred_bootstrap_rerun/"))+1
    importance_matrix %>% fwrite(paste0("../test/dimred_bootstrap_rerun/Dimred_run2019toPscoreMean_",thisone,".csv",sep=""))
    finalmodel_dimred$results
}
)


#fwrite(result_bootstrap_dimred,file = "../results/06_dimred_19_to_pscoreMean_bootstrap_1000.csv")
```

```{r}
Xgboostinput_first<-merge(table2019_wdiImputed,excessCounrty_monthly_pvalue %>% select(p_value_mort,iso3),by.x = "Country Code",by.y="iso3")


data_wdi<-Xgboostinput_first
trials <- 1
cl <- makeCluster(20)
registerDoParallel(cl)
# Evaluate the performance of the model using RMSE


system.time(result_bootstrap <- foreach(icount(trials),
                  .combine = bind_rows) %dopar% {
                    require(xgboost)
                    require(tidyverse)
                    require(data.table)
    train_index <- sample(1:nrow(data_wdi), 0.8 * nrow(data_wdi))
    features<-colnames(data_wdi)[!grepl("Country|year|p_value|pscore_mean|[1-9]{2}",colnames(data_wdi))]

    Label <- "p_value_mort"
    trainDAT <- data_wdi[train_index, ]
    testDAT <- data_wdi[-train_index, ]
    test_proportion <- 0.8
    

    
    # Train the XGBoost model using the function
    trainedModel<- train_xgboost_model(trainDAT, Label, features, custom_control, param_grid)
    # Assuming you have the 'finalmodel' and 'test_data' from the previous steps
    predictedValue<-postResample(
      predict(trainedModel, 
          testDAT %>% dplyr::select(features)),  
        testDAT %>% pull(Label)
      )
    predictedValue
    
    # xgb_caret <- caret::train(x = trainDAT %>% dplyr::select(all_of(features)),
    #                y = trainDAT %>% pull(label),
    #                method = "xgbTree",
    #                verbosity = 0,
    #                trControl=custom_control,
    #   tuneGrid = param_grid
    #                #,tuneLength = 2
    # 
    #                 )
    # 
    # finalmodel <- caret::train(x = trainDAT %>% dplyr::select(all_of(features)),
    #                y = trainDAT %>% pull(label),
    #                method = "xgbTree",
    #                verbosity = 0,
    #                ## CAST CV LLO (Hanna Meyer)
    #                trControl=trainControl(method="cv",
    #                                       number = 30),
    # 
    #                tuneGrid = xgb_caret$bestTune)
    # 
    # importance_matrix<-varImp(finalmodel)$importance %>%
    #      mutate(var = rownames(varImp(xgb_caret)$importance))%>%
    #      dplyr::rename(importance = Overall) 
   
    

    
}
)


result_bootstrap_dimred_2019toPScoreMean<-result_bootstrap
fwrite(result_bootstrap_dimred_2019toPScoreMean,file = "../results/07_wdi_19_to_pscoreMean_bootstrap_1000.csv")
# load("../data/06_wdi_20_bootstrap_1000")


```

### plot both
```{r}
a<-fread("../results/06_wdi_19_to_pscoreMean_bootstrap_1000.csv")
b<-fread("../results/06_dimred_19_to_pscoreMean_bootstrap_1000.csv")
a$Dataset<-"WDI imputed"
b$Dataset<-"Latent Space"
plotR<-rbind(a,b) %>% ggplot(aes(x=Rsquared,fill=Dataset))+geom_histogram(bins = 100)+theme_classic()+xlim(0,1)+scale_fill_manual(values = col_pal)+theme(legend.position = "sad")+ylab("N")+xlab(expression(paste("R"^2)))
plotRMS<-rbind(a,b) %>% ggplot(aes(x=RMSE,fill=Dataset))+geom_histogram(bins = 100)+theme_classic()+xlim(5,15)+scale_fill_manual(values = col_pal)+theme(legend.position = "sad")+ylab("")

plot_bootstrap<-ggarrange(plotR,plotRMS)
plot_bootstrap_legend<- rbind(a,b) %>% ggplot(aes(x=RMSE,fill=Dataset))+geom_histogram(bins = 100)+scale_fill_manual(values = col_pal)

plot_bootstrap%>%  ggsave(file="../figures/06_plots/c_compare_prediction_bootstrap.png",device = "png",width = 9,height = 4)
plot_bootstrap_legend%>%  ggsave(file="../figures/06_plots/c_compare_prediction_bootstrap_legend.png",device = "png")

```
# Result 3 Monthly excess mort
## first round 2019 dimred
```{r}
features<-c("iso1","iso2","iso3","iso4","iso5","iso6","iso7")
dimredData<-dimred2021_knn250_2[] %>% filter(year%in% 2019) %>% select(c(`Country Code`,features)) %>%  filter(!is.na(iso1))
excessmonthly2<-merge(excessmonthly %>% select(date,`Country Code`,p_value_mort),dimredData,by="Country Code")
listofmonth<-excessmonthly$date %>%  unique 
listofmonth<-listofmonth[!is.na(listofmonth)] 
require(xgboost)
require(tidyverse)
require(data.table)
library(shapper)
require("SHAPforxgboost")
cooloutput<-data.frame()


features<-c("iso1","iso2","iso3","iso4","iso5","iso6","iso7")
     label<-"p_value_mort"

     
     system.time(result_bootstrap_dimred <- for(i in 1:length(listofmonth)){
    


    
    
    
    thismonth<-listofmonth[i]
    excessmonthly2 %>%  filter(date%in% thismonth) %>% select(all_of(c(features,label)))-> runthismonthlysubset
   
    trainIndex <- createDataPartition(runthismonthlysubset$p_value_mort, p=0.8, list=FALSE)
    train_df <- runthismonthlysubset[trainIndex,]
    test_df <- runthismonthlysubset[-trainIndex,]

    a<-train_df %>% dplyr::select(all_of(features)) %>% as.matrix()
    xgb_caret_dimred <- caret::train(x = train_df %>% dplyr::select(all_of(features)),
                   y = train_df %>% pull(label),
                   method = "xgbTree",
                   verbosity = 0,trControl=custom_control,tuneGrid = param_grid
                   ,tuneLength = 10
    
                    )
    
    # xgb_caret_dimred
    # mod <- xgboost::xgboost(data = a, 
    #                     label = runthismonthlysubset$p_value_mort, 
    #                     params = param_list, nrounds = 10,
    #                     verbose = FALSE,
    #                     early_stopping_rounds = 8)
    
    finalmodel_dimred <- caret::train(x = train_df %>% dplyr::select(all_of(features)),
                   y = train_df %>% pull(label),
                   method = "xgbTree",
                   verbosity = 0,
                   ## CAST CV LLO (Hanna Meyer)
                   trControl=trainControl(method="cv",
                                          number = 60),
                   # early_stopping_rounds = 10,
                   tuneGrid = xgb_caret_dimred$bestTune)
    
    importance_matrix<-varImp(finalmodel_dimred)$importance %>%
         mutate(var = rownames(varImp(xgb_caret_dimred)$importance))%>%
         dplyr::rename(importance = Overall) 
    importance_matrix$name<- thismonth
    importance_matrix$rsquared<-finalmodel_dimred$results$Rsquared
    predictedValue<-postResample(
      predict(finalmodel_dimred, 
          test_df %>% dplyr::select(-label)),  
        test_df %>% pull(label)
      )
    
    shap_values <- shap.values(xgb_model = finalmodel_dimred$finalModel, X_train = as.matrix(train_df %>% dplyr::select(all_of(features))))
    shap_values
    # The ranked features by mean |SHAP|
    shaplmean<-data.frame(Shapl_feature=shap_values$mean_shap_score) %>% rownames_as_column(var="var")
    
    
      
    importance_matrix$rsquared_tested<-predictedValue[2]
    importance_matrix$rmse_tested<-predictedValue[1]
    importance_matrix<-cbind(importance_matrix,shaplmean[match(shaplmean$var,importance_matrix$var),])
    # shap_values <- shap.values(xgb_model = mod, X_train = a)
    # # The ranked features by mean |SHAP|
    # shaploutput<-shap_values$mean_shap_score 
    # 
    # orderShap<-match(rownames(importance_matrix), names(shaploutput))
    # 
    # importance_matrix$shapmean<-shaploutput[orderShap]
    cooloutput<-rbind(cooloutput,importance_matrix)
    
})

```
```{r}

rsquared_values_plot<-
  cooloutput %>% select(name,rsquared_tested) %>%  unique %>% ggplot( . , aes(x=name, y=rsquared_tested)) +
  geom_segment( aes(x=name, xend=name, y=0, yend=rsquared_tested), color="grey") +
  geom_point( color=col_pal[5], size=4) +
  theme_light() + geom_line()+
  theme(
    panel.grid.major.x = element_blank(),
    panel.border = element_blank(),
    axis.ticks.x = element_blank()
  ) +
  xlab("") +
  ylab("Value of Y") +ylim(0.2,1)
# heatmap
rsquared_values_plot
```


```{r}


start_date<- as.Date("2020-1-1")
end_date<- as.Date("2021-12-1")
# Heatmap 
cooloutput$var1<-cooloutput$var
cooloutput$var<-NULL
excessmonthy <- excessCounrty_monthly_pvalue %>% arrange(country,year,month) %>% rename(`Country Code`=iso3) %>% 
  mutate(date = as.Date(paste(year, month, "01", sep = "-"), format = "%Y-%m-%d"),identifier=paste0(`Country Code`,"_",year(date),"_",month(date)))%>% filter(date>"2020-02-01") %>% merge( . , thisisOWID,by.x="identifier",by.y="monthyear")
excessmonthy <- excessCounrty_monthly_pvalue %>% arrange(country,year,month) %>% rename(`Country Code`=iso3) %>% 
  mutate(date = as.Date(paste(year, month, "01", sep = "-"), format = "%Y-%m-%d"),identifier=paste0(`Country Code`,"_",year(date),"_",month(date)))%>% filter(date>"2020-02-01") %>% merge( . , thisisOWID,by.x="identifier",by.y="monthyear")
importanceplot<-cooloutput %>% 
  mutate(var = factor(var, levels=rev(features))) %>% 
  ggplot( . , aes(name, var, fill= Shapl_feature)) + 
  geom_tile()+scale_fill_viridis()+theme_classic()+xlab("")+xlim(1,1)+
  scale_x_date(limits = c(start_date, end_date))+
  theme(axis.text.x = element_text(angle = 90, hjust = 1))+theme(legend.position = "sad")

plot_grid(rsquared_values_plot,exmortelement, importanceplot,ncol = 1,rel_heights =     c(1.5/5, 2/5,3/5))

```

```{r}
postResample(
  predict(finalmodel, 
          test_df %>% dplyr::select(-share_exposed_pop)),  
  test_df %>% pull(share_exposed_pop)
  )

## nash-sutville efficiency
NSE(predict(finalmodel, 
          test_df %>% dplyr::select(-share_exposed_pop)), 
    test_df %>% pull(share_exposed_pop))
```


## second round 2020 and 2021

### looks horrible dont do it again
```{r}

features<-c("iso1","iso2","iso3","iso4","iso5","iso6","iso7")
dimredData2020<-dimred2021_knn250_2[] %>% filter(year%in% 2020) %>% 
  mutate(identifier=paste0(`Country Code`,"_",year)) %>% 
  select(c(identifier,features)) %>%  filter(!is.na(iso1)) 
dimredData2021<-dimred2021_knn250_2[] %>% filter(year%in% 2021) %>% 
  mutate(identifier=paste0(`Country Code`,"_",year)) %>% 
  select(c(identifier,features)) %>%  filter(!is.na(iso1)) 
dimredDataboth<-rbind(dimredData2020,dimredData2021)
excessmonthly<-excessmonthly %>% 
  mutate(identifier=paste0(`Country Code`,"_",year))
excessmonthly2<-merge(excessmonthly,dimredDataboth,by="identifier")
listofmonth<-excessmonthly$date %>%  unique 
listofmonth<-listofmonth[!is.na(listofmonth)] 
require(xgboost)
require(tidyverse)
require(data.table)
library(shapper)
require("SHAPforxgboost")
cooloutput_2020_2021<-data.frame()

system.time(result_bootstrap_dimred <- for(i in 1:length(listofmonth)){
  



    
    features<-c("iso1","iso2","iso3","iso4","iso5","iso6","iso7")
    thismonth<-listofmonth[i]
    excessmonthly2 %>%  filter(date%in% thismonth) -> runthismonthlysubset
    label<-"pscore_mean"
    
    a<-runthismonthlysubset %>% dplyr::select(all_of(features)) %>% as.matrix()
    xgb_caret_dimred <- caret::train(x = runthismonthlysubset %>% dplyr::select(all_of(features)),
                   y = runthismonthlysubset %>% pull("p_value_mort"),
                   method = "xgbTree",
                   verbosity = 0,trControl=custom_control,tuneGrid = param_grid
                   ,tuneLength = 10
    
                    )
    
    # xgb_caret_dimred
    # mod <- xgboost::xgboost(data = a, 
    #                     label = runthismonthlysubset$p_value_mort, 
    #                     params = param_list, nrounds = 10,
    #                     verbose = FALSE,
    #                     early_stopping_rounds = 8)
    
    finalmodel_dimred <- caret::train(x = runthismonthlysubset %>% dplyr::select(all_of(features)),
                   y = runthismonthlysubset %>% pull(p_value_mort),
                   method = "xgbTree",
                   verbosity = 0,
                   ## CAST CV LLO (Hanna Meyer)
                   trControl=trainControl(method="cv",
                                          number = 30),
                   # early_stopping_rounds = 10,
                   tuneGrid = xgb_caret_dimred$bestTune)
    
    importance_matrix<-varImp(finalmodel_dimred)$importance %>%
         mutate(var = rownames(varImp(xgb_caret_dimred)$importance))%>%
         dplyr::rename(importance = Overall) 
    importance_matrix$name<- thismonth
    importance_matrix$rsquared<-finalmodel_dimred$results$Rsquared
    
    # shap_values <- shap.values(xgb_model = mod, X_train = a)
    # # The ranked features by mean |SHAP|
    # shaploutput<-shap_values$mean_shap_score 
    # 
    # orderShap<-match(rownames(importance_matrix), names(shaploutput))
    # 
    # importance_matrix$shapmean<-shaploutput[orderShap]
    cooloutput_2020_2021<-rbind(cooloutput_2020_2021,importance_matrix)
    
})
```
```{r}

rsquared_values_plot<-
  cooloutput_2020_2021 %>% select(name,rsquared) %>%  unique %>% ggplot( . , aes(x=name, y=rsquared)) +
  geom_segment( aes(x=name, xend=name, y=0, yend=rsquared), color="grey") +
  geom_point( color=col_pal[5], size=4) +
  theme_light() + geom_line()+
  theme(
    panel.grid.major.x = element_blank(),
    panel.border = element_blank(),
    axis.ticks.x = element_blank()
  ) +
  xlab("") +
  ylab("Value of Y") +ylim(0.2,1)
# heatmap
```


```{r}


start_date<- as.Date("2020-1-1")
end_date<- as.Date("2021-12-1")
# Heatmap 
importanceplot<-cooloutput_2020_2021 %>% filter(!is.na(var))%>% 
  mutate(var = factor(var, levels=rev(features))) %>% 
  ggplot( . , aes(name, var, fill= importance)) + 
  geom_tile()+scale_fill_viridis()+theme_classic()+xlab("")+xlim(1,1)+
  scale_x_date(limits = c(start_date, end_date))+
  theme(axis.text.x = element_text(angle = 90, hjust = 1))+theme(legend.position = "sad")

plot_grid(rsquared_values_plot,exmortelement, importanceplot,ncol = 1,rel_heights =     c(1.5/5, 2/5,3/5))

```
## third round 2019 dimred on clustes
### the clusters
```{r}
clusterbyExcessandIsos<-dimred2021_knn250_2 %>% filter(year%in% 2019,`Country Code`!="PER") %>%  select(`Country Code`,pscore_mean,features) %>%  drop_na()
kmeans_result <- kmeans(clusterbyExcessandIsos[,-1], centers = 5)
clusterbyExcessandIsos$Cluster<-kmeans_result$cluster
```

```{r}
theclusteredFrame<-data.frame()
for(i in unique(clusterbyExcessandIsos$Cluster))
  
  
  

var_caountry<-clusterbyExcessandIsos[clusterbyExcessandIsos$Cluster %in% i,] %>% select(`Country Code`) %>%  unlist()
features<-c("iso1","iso2","iso3","iso4","iso5","iso6","iso7")
dimredData<-dimred2021_knn250_2[] %>% filter(year%in% 2019) %>% filter(`Country Code`%in% var_caountry) %>% select(c(`Country Code`,features)) %>%  filter(!is.na(iso1))
excessmonthly2<-merge(excessmonthly %>% select(date,`Country Code`,p_value_mort),dimredData,by="Country Code")

listofmonth<-excessmonthly$date %>%  unique 
listofmonth<-listofmonth[!is.na(listofmonth)] 
require(xgboost)
require(tidyverse)
require(data.table)
library(shapper)
require("SHAPforxgboost")
cooloutput<-data.frame()

system.time(result_bootstrap_dimred <- for(i in 1:length(listofmonth)){
  

    
    features<-c("iso1","iso2","iso3","iso4","iso5","iso6","iso7")
    thismonth<-listofmonth[i]
    excessmonthly2 %>%  filter(date%in% thismonth) -> runthismonthlysubset
    label<-"pscore_mean"
    
    a<-runthismonthlysubset %>% dplyr::select(all_of(features)) %>% as.matrix()
    xgb_caret_dimred <- caret::train(x = runthismonthlysubset %>% dplyr::select(all_of(features)),
                   y = runthismonthlysubset %>% pull("p_value_mort"),
                   method = "xgbTree",
                   verbosity = 0,trControl=custom_control,tuneGrid = param_grid
                   ,tuneLength = 10
    
                    )
    
    # xgb_caret_dimred
    # mod <- xgboost::xgboost(data = a, 
    #                     label = runthismonthlysubset$p_value_mort, 
    #                     params = param_list, nrounds = 10,
    #                     verbose = FALSE,
    #                     early_stopping_rounds = 8)
    
    finalmodel_dimred <- caret::train(x = runthismonthlysubset %>% dplyr::select(all_of(features)),
                   y = runthismonthlysubset %>% pull(p_value_mort),
                   method = "xgbTree",
                   verbosity = 0,
                   ## CAST CV LLO (Hanna Meyer)
                   trControl=trainControl(method="cv",
                                          number = 60),
                   # early_stopping_rounds = 10,
                   tuneGrid = xgb_caret_dimred$bestTune)
    
    importance_matrix<-varImp(finalmodel_dimred)$importance %>%
         mutate(var = rownames(varImp(xgb_caret_dimred)$importance))%>%
         dplyr::rename(importance = Overall) 
    importance_matrix$name<- thismonth
    importance_matrix$rsquared<-finalmodel_dimred$results$Rsquared
    
    # shap_values <- shap.values(xgb_model = mod, X_train = a)
    # # The ranked features by mean |SHAP|
    # shaploutput<-shap_values$mean_shap_score 
    # 
    # orderShap<-match(rownames(importance_matrix), names(shaploutput))
    # 
    # importance_matrix$shapmean<-shaploutput[orderShap]
    cooloutput<-rbind(cooloutput,importance_matrix)
    
})
cooloutput$cluster<-i
theclusteredFrame<-bind_rows(theclusteredFrame,cooloutput)


```
```{r}

rsquared_values_plot<-
  cooloutput %>% select(name,rsquared) %>%  unique %>% ggplot( . , aes(x=name, y=rsquared)) +
  geom_segment( aes(x=name, xend=name, y=0, yend=rsquared), color="grey") +
  geom_point( color=col_pal[5], size=4) +
  theme_light() + geom_line()+
  theme(
    panel.grid.major.x = element_blank(),
    panel.border = element_blank(),
    axis.ticks.x = element_blank()
  ) +
  xlab("") +
  ylab("Value of Y") +ylim(0.2,1)
# heatmap
```


```{r}


start_date<- as.Date("2020-1-1")
end_date<- as.Date("2021-12-1")
# Heatmap 
importanceplot<-cooloutput %>% filter(!is.na(var))%>% 
  mutate(var = factor(var, levels=rev(features))) %>% 
  ggplot( . , aes(name, var, fill= importance)) + 
  geom_tile()+scale_fill_viridis()+theme_classic()+xlab("")+xlim(1,1)+
  scale_x_date(limits = c(start_date, end_date))+
  theme(axis.text.x = element_text(angle = 90, hjust = 1))+theme(legend.position = "sad")

plot_grid(rsquared_values_plot,exmortelement, importanceplot,ncol = 1,rel_heights =     c(1.5/5, 2/5,3/5))

```

# Results 5 circular plot Signficant correlated WDI features in a 
`
## PLOT 1 all data
```{r}
DATA_DIR<-"../data/"
RELATIVE_INDICATORS_FILE_NAME <- "../data/01_data/relative_indicators_2023_06.R"

order_indidactors<-iso1_onlyDimred$`Series Code` 
dcor_var_iso_cons<-matrix(nrow =length(order_indidactors),ncol=7)
rownames(dcor_var_iso_cons)<-order_indidactors
dcor_var_iso_cons[,1]<-iso1_onlyDimred[iso1_onlyDimred$`Series Code`%in% order_indidactors,"iso1_all"] %>%  unlist
dcor_var_iso_cons[,2]<-iso2_onlyDimred[iso2_onlyDimred$`Series Code`%in% order_indidactors,"iso2_all"] %>%  unlist
dcor_var_iso_cons[,3]<-iso3_onlyDimred[iso3_onlyDimred$`Series Code`%in% order_indidactors,"iso3_all"] %>%  unlist
dcor_var_iso_cons[,4]<-iso4_onlyDimred[iso4_onlyDimred$`Series Code`%in% order_indidactors,"iso4_all"] %>%  unlist
dcor_var_iso_cons[,5]<-iso5_onlyDimred[iso5_onlyDimred$`Series Code`%in% order_indidactors,"iso5_all"] %>%  unlist
dcor_var_iso_cons[,6]<-iso6_onlyDimred[iso6_onlyDimred$`Series Code`%in% order_indidactors,"iso6_all"] %>%  unlist
dcor_var_iso_cons[,7]<-iso7_onlyDimred[iso7_onlyDimred$`Series Code`%in% order_indidactors,"iso7_all"] %>%  unlist


used_var_names_max_dcor <- c(
  dcor_var_iso_cons[, 1] %>% order(decreasing = TRUE) %>% { .[1:15] } %>% { names(dcor_var_iso_cons[., 1]) },
  dcor_var_iso_cons[, 2] %>% order(decreasing = TRUE) %>% { .[1:15] } %>% { names(dcor_var_iso_cons[., 2]) },
  dcor_var_iso_cons[, 3] %>% order(decreasing = TRUE) %>% { .[1:30] } %>% { names(dcor_var_iso_cons[., 3]) },
  dcor_var_iso_cons[, 4] %>% order(decreasing = TRUE) %>% { .[1:30] } %>% { names(dcor_var_iso_cons[., 4]) },
  dcor_var_iso_cons[, 5] %>% order(decreasing = TRUE) %>% { .[1:05] } %>% { names(dcor_var_iso_cons[., 5]) },
  dcor_var_iso_cons[, 6] %>% order(decreasing = TRUE) %>% { .[1:05] } %>% { names(dcor_var_iso_cons[., 6]) },
  dcor_var_iso_cons[, 7] %>% order(decreasing = TRUE) %>% { .[1:10]  } %>% {names(dcor_var_iso_cons[., 7]) }
) %>% unique %>% sort
subsetwdi<-wdi_explained[wdi_explained$`Series Code`%in% used_var_names_max_dcor,]

```

#### input
```{r}
source("../libraries/01a_lib_social_indicators.R")
DATA_DIR<-"../data/"
RELATIVE_INDICATORS_FILE_NAME <- "../data/01_data/relative_indicators_2023_06.R"
var_names         = used_var_names_max_dcor
dcor_var_iso_cons = dcor_var_iso_cons
load("../data/wdi_series")
local_wdi_series  = wdi_series[,]



```

#### setup
```{r}




subsetwdi<-wdi_explained[wdi_explained$`Series Code`%in% used_var_names_max_dcor,]

var_names         = used_var_names_max_dcor
dcor_var_iso_cons = dcor_var_iso_cons
load("../data/wdi_series")
local_wdi_series  = wdi_series[,]
expand_multiplier = 1 / 10
text_size         = 5
label_modifier    = manual_label_modifier
col_pal <- c("#1F968B", "#3CBB75", "#7570b3", "#e7298a", "yellow3", "orange","slategrey")
```

### plot function
```{r}

#label_modifier    = short_name_label_modifier,
## .overwrite = TRUE, # nolint: commented_code_linter.
.width  = 15
.height = 15

component_common <- "Dimension"
component_names <- paste(component_common, 1:7)
xy_expand <- function(xy, expand) xy * (1 + expand * expand_multiplier)
# Does not work
local_wdi_series<-local_wdi_series[grepl(":",local_wdi_series$Topic),]

local_wdi_series <- wdi_series_split_topic(local_wdi_series)

wdi_series_used_important <- local_wdi_series[`Indicator Code` %in% var_names]
var_names<-var_names[var_names%in% wdi_series_used_important$`Indicator Code`]
require(igraph)
stopifnot(length(var_names) == nrow(wdi_series_used_important))
dcor_var_iso_cons_subset<-dcor_var_iso_cons[rownames(dcor_var_iso_cons)%in% var_names,]
dcor_var_iso_cons<-dcor_var_iso_cons[rownames(dcor_var_iso_cons)%in% var_names,]



local_wdi_series<-local_wdi_series[local_wdi_series$`Indicator Code`%in% var_names,]
require(tidygraph)
require(ggraph)

dcor_subset<-dcor_var_iso_cons_subset
## local_wdi_series$`Indicator Code`
edges_level_01 <- local_wdi_series %>% select(`Indicator Code`, Topic1) %>%
  unique %>% rename( from = Topic1, to = `Indicator Code`)
edges_level_12 <- local_wdi_series %>% select(Topic1, Topic2)           %>%
  unique %>% rename( from = Topic2, to = Topic1          )
edges_level_23 <- local_wdi_series %>% select(Topic2, Topic3)           %>%
  unique %>% rename( from = Topic3, to = Topic2          )
edges_level_34 <- local_wdi_series %>% select(Topic3, Topic4)           %>%
  unique %>% rename( from = Topic4, to = Topic3          )
edges_level_45 <- local_wdi_series %>% select(Topic4, Topic5)           %>%
  unique %>% rename( from = Topic5, to = Topic4          )
## create the actual igraph object

rbind(edges_level_01, edges_level_12, edges_level_23, edges_level_34,
      edges_level_45) %>%
unique %>%
filter(to != from) ->listi
hier_graph <-listi %>% data.frame() %>%  { .[rev(seq_len(ncol(.)))] } %>% graph_from_data_frame
  ## graph_from_data_frame does not care about column names,
  ## only about positions
  # %>%
  #graph_from_data_frame

## additional data:

graph_idxs <- match(var_names, V(hier_graph)$name)

hier_graph <- set_vertex_attr(hier_graph, "which_max_iso", graph_idxs,
                              dcor_subset %>% apply(1, which.max))
hier_graph <- set_vertex_attr(hier_graph, "dcor_max", graph_idxs,
                              dcor_subset %>% apply(1, max))
hier_graph <- set_vertex_attr(hier_graph, "dcor1", graph_idxs,
                              dcor_subset[, 1])
hier_graph <- set_vertex_attr(hier_graph, "dcor2", graph_idxs,
                              dcor_subset[, 2])
hier_graph <- set_vertex_attr(hier_graph, "dcor3", graph_idxs,
                              dcor_subset[, 3])
hier_graph <- set_vertex_attr(hier_graph, "dcor4", graph_idxs,
                              dcor_subset[, 4])
hier_graph <- set_vertex_attr(hier_graph, "dcor5", graph_idxs,
                              dcor_subset[, 5])
hier_graph <- set_vertex_attr(hier_graph, "dcor6", graph_idxs,
                              dcor_subset[, 6])
hier_graph <- set_vertex_attr(hier_graph, "dcor7", graph_idxs,
                              dcor_subset[, 7])
# 

hier_graph <-
  set_vertex_attr(hier_graph, "Indicator.Code", graph_idxs,
                  local_wdi_series$`Indicator Code`[
                               match(var_names, local_wdi_series$`Indicator Code`)])



# thisPeople<-fread("../data/indicator_compressed_names.csv")
# wdi_names_prepared<-wdi_series_used_important[,c("Indicator Code","Indicator Name")]
# 
# 
# merge(wdi_names_prepared,thisPeople,by="Indicator Code",all.x = T)->wdi_names_prepared2

# wdi_names_prepared2 %>% select(`Indicator Code`,`Indicator Name`,`Compressed Name`)%>%  filter(is.na(`Compressed Name`)) %>% fwrite("../output/preparethenameschatgpt.csv")
# 
# wdi_series_used[,c("Indicator Code","Compressed Name")] #%>% fwrite("../data/indicator_compressed_names.csv")
# # # chatgpt shortened the names
# wdi_series_used
# wdi_series_used2<-fread("../data/indicator_compressed_names.csv")
# shortnames<-fread("../data/05_abbrevationmaker.csv",sep=";")
# rbind(wdi_series_used2[!(wdi_series_used2$`Indicator Code`%in% shortnames$`Indicator Code`)],shortnames) %>%  fwrite("../output/05_indicator_compressed_names.csv")

wdi_series_used<-fread("../data/indicator_compressed_names.csv")

wdi_var_hierarchy_graph <-hier_graph
axis_cons <- lapply(1:7, inner_connections_by_axis,
                       dcor_var_iso_cons       = dcor_var_iso_cons,
                       wdi_var_hierarchy_graph = wdi_var_hierarchy_graph,
                       var_names               = var_names)




expand_multiplier = 1/2
text_size         = 3
label_modifier    = manual_label_modifier
#label_modifier    = short_name_label_modifier
## .overwrite = TRUE, # nolint: commented_code_linter.
.width  = 50
.height = 50


wdi_var_hierarchy_graph %>%
ggraph(layout = "dendrogram", circular = TRUE) +
geom_node_point(aes(filter = leaf,
                    x = xy_expand(x, dcor1),
                    y = xy_expand(y, dcor1)),
                colour = col_pal[1]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor1),
                 y = xy_expand(y, dcor1)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[1]) +
geom_node_point(aes(filter = leaf,
                   x = xy_expand(x, dcor2),
                    y = xy_expand(y, dcor2)),
                colour = col_pal[2]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor2),
                 y = xy_expand(y, dcor2)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[2]) +
geom_node_point(aes(filter = leaf,
                    x = xy_expand(x, dcor3),
                    y = xy_expand(y, dcor3)),
                colour = col_pal[3]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor3),
                 y = xy_expand(y, dcor3)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[3]) +
geom_node_point(aes(filter = leaf,
                    x = xy_expand(x, dcor4),
                    y = xy_expand(y, dcor4)),
                colour = col_pal[4]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor4),
                 y = xy_expand(y, dcor4)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[4]) +
geom_node_point(aes(filter = leaf,
                    x = xy_expand(x, dcor5),
                    y = xy_expand(y, dcor5)),
                colour = col_pal[5]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor5),
                 y = xy_expand(y, dcor5)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[5]) +
  geom_node_point(aes(filter = leaf,
                    x = xy_expand(x, dcor5),
                    y = xy_expand(y, dcor5)),
                colour = col_pal[5]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor5),
                 y = xy_expand(y, dcor5)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[5]) +
  geom_node_point(aes(filter = leaf,
                    x = xy_expand(x, dcor5),
                    y = xy_expand(y, dcor5)),
                colour = col_pal[6]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor5),
                 y = xy_expand(y, dcor5)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[6]) +
      geom_node_point(aes(filter = leaf,
                    x = xy_expand(x, dcor5),
                    y = xy_expand(y, dcor5)),
                colour = col_pal[7]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor5),
                 y = xy_expand(y, dcor5)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[7]) +
geom_node_point(aes(filter = leaf, color = factor(which_max_iso)),
                size = 0)+
geom_node_text(aes(filter = leaf,
                   label  = label_modifier(as.character(Indicator.Code)),
                   color  = factor(paste(component_common, which_max_iso)),
                   x      = x * (1 + 1 * expand_multiplier),
                   y      = y * (1 + 1 * expand_multiplier),
                   hjust  = if_else(x > 0, 0, 1),
                   angle  = atan(y / x) / 2 / pi * 360),
               show.legend = FALSE,
               size        = text_size) +
geom_conn_bundle(data = get_con(axis_cons[[1]][1, ], axis_cons[[1]][2, ]),
                 edge_colour = col_pal[1], edge_alpha = 0.3) +
geom_conn_bundle(data = get_con(axis_cons[[2]][1, ], axis_cons[[2]][2, ]),
                 edge_colour = col_pal[2], edge_alpha = 0.3) +
geom_conn_bundle(data = get_con(axis_cons[[3]][1, ], axis_cons[[3]][2, ]),
                 edge_colour = col_pal[3], edge_alpha = 0.3) +
(
  if (ncol(axis_cons[[4]]) > 0)
    geom_conn_bundle(data = get_con(axis_cons[[4]][1, ],
                                    axis_cons[[4]][2, ]),
                     edge_colour = col_pal[4], edge_alpha = 0.3)
  else
    list()
) +
geom_conn_bundle(data = get_con(axis_cons[[5]][1, ], axis_cons[[5]][2, ]),
                 edge_colour = col_pal[5], edge_alpha = 0.3) +
geom_edge_diagonal() +
# geom_node_label(aes(filter = !leaf,
#                     fontface = if_else(name %in% level1, "bold", "plain"),
#                     label = name, angle = atan(y / x) / 2 / pi * 360),
#                 size = 2) +
scale_color_manual(limits = component_names,
                   labels = component_names,
                   values = col_pal) +
guides(colour = guide_legend(title = component_common,
                             override.aes = list(size = 5))) +
theme_void() +
theme(legend.title = element_blank(),
      legend.position = c(0.9, 0.1),
      legend.justification = c(0.9, 0.1)) +
expand_limits(x = c(-2.7, 2.7), y = c(-2.7, 2.7))
# 
ggsave("../figures/07_230808_circle_all.pdf",device = "pdf",height = 12,width = 12)

```
#### Description
```{r}
subsetwdi$Topic %>%  table %>%  data.frame() %>%  arrange(desc(Freq))
```

## PLOT 2 SUBSET Significant shapleys


### input
#### data

```{r}
DATA_DIR<-"../data/"
RELATIVE_INDICATORS_FILE_NAME <- "../data/01_data/relative_indicators_2023_06.R"

order_indidactors<-iso1_onlyDimred$`Series Code` 
dcor_var_iso_cons<-matrix(nrow =length(order_indidactors),ncol=7)
rownames(dcor_var_iso_cons)<-order_indidactors

iso1_onlyDimred %>% filter(N>=mean(N))->iso1_onlyDimred_2
iso2_onlyDimred %>% filter(N_shapley>=mean(N_shapley))->iso2_onlyDimred_2
iso3_onlyDimred %>% filter(N_shapley>=mean(N_shapley))->iso3_onlyDimred_2
iso4_onlyDimred %>% filter(N_shapley>=mean(N_shapley))->iso4_onlyDimred_2
iso5_onlyDimred %>% filter(N_shapley>=mean(N_shapley))->iso5_onlyDimred_2
iso6_onlyDimred %>% filter(N_shapley>=mean(N_shapley))->iso6_onlyDimred_2
iso7_onlyDimred %>% filter(N_shapley>=mean(N_shapley))->iso7_onlyDimred_2

theseindicators<-unique(c(iso1_onlyDimred_2$`Series Code`,iso2_onlyDimred_2$`Series Code`,iso3_onlyDimred_2$`Series Code`,
                          iso4_onlyDimred_2$`Series Code`,iso5_onlyDimred_2$`Series Code`,iso6_onlyDimred_2$`Series Code`,
                          iso7_onlyDimred_2$`Series Code`))

dcor_var_iso_cons[,1]<-iso1_onlyDimred[iso1_onlyDimred$`Series Code`%in% order_indidactors,"iso1"] %>%  unlist
dcor_var_iso_cons[,2]<-iso2_onlyDimred[iso2_onlyDimred$`Series Code`%in% order_indidactors,"iso2_75to100_75to100"] %>%  unlist
dcor_var_iso_cons[,3]<-iso3_onlyDimred[iso3_onlyDimred$`Series Code`%in% order_indidactors,"iso3_shapley"] %>%  unlist
dcor_var_iso_cons[,4]<-iso4_onlyDimred[iso4_onlyDimred$`Series Code`%in% order_indidactors,"iso4_0to25_0to25"] %>%  unlist
dcor_var_iso_cons[,5]<-iso5_onlyDimred[iso5_onlyDimred$`Series Code`%in% order_indidactors,"iso5_shapley"] %>%  unlist
dcor_var_iso_cons[,6]<-iso6_onlyDimred[iso6_onlyDimred$`Series Code`%in% order_indidactors,"iso6_shapley"] %>%  unlist
dcor_var_iso_cons[,7]<-iso7_onlyDimred[iso7_onlyDimred$`Series Code`%in% order_indidactors,"iso7_shapley"] %>%  unlist
dcor_var_iso_cons<-dcor_var_iso_cons[rownames(dcor_var_iso_cons) %in%theseindicators,]
dcor_var_iso_cons %>% 
```

#### setup
```{r}


used_var_names_max_dcor <- c(
  dcor_var_iso_cons[, 1] %>% order(decreasing = TRUE) %>% { .[1:15] } %>% { names(dcor_var_iso_cons[., 1]) },
  dcor_var_iso_cons[, 2] %>% order(decreasing = TRUE) %>% { .[1:20] } %>% { names(dcor_var_iso_cons[., 2]) },
  dcor_var_iso_cons[, 3] %>% order(decreasing = TRUE) %>% { .[1:30] } %>% { names(dcor_var_iso_cons[., 3]) },
  dcor_var_iso_cons[, 4] %>% order(decreasing = TRUE) %>% { .[1:30] } %>% { names(dcor_var_iso_cons[., 4]) },
  dcor_var_iso_cons[, 5] %>% order(decreasing = TRUE) %>% { .[1:05] } %>% { names(dcor_var_iso_cons[., 5]) },
  dcor_var_iso_cons[, 6] %>% order(decreasing = TRUE) %>% { .[1:05] } %>% { names(dcor_var_iso_cons[., 6]) },
  dcor_var_iso_cons[, 7] %>% order(decreasing = TRUE) %>% { .[1:05]  } %>% {names(dcor_var_iso_cons[., 7]) }
) %>% unique %>% sort

subsetwdi<-wdi_explained[wdi_explained$`Series Code`%in% used_var_names_max_dcor,]

var_names         = used_var_names_max_dcor
dcor_var_iso_cons = dcor_var_iso_cons
load("../data/wdi_series")
local_wdi_series  = wdi_series[,]
expand_multiplier = 1 / 10
text_size         = 5
label_modifier    = manual_label_modifier
col_pal <- c("#1F968B", "#3CBB75", "#7570b3", "#e7298a", "yellow3", "orange","slategrey")
```

#### modifiy the compressed names
```{r}
compressedNames <- fread("../data/indicator_compressed_names.csv")

missingIndicators <- wdi_explained %>%
  filter(`Series Code` %in% used_var_names_max_dcor) %>%
  select(`Series Code`,`Indicator Name`) %>%
  rename(`Indicator Code` = `Series Code`) %>%
  merge(compressedNames[,c("Indicator Code","Compressed Name")], by = "Indicator Code", all.x = TRUE) %>%
  filter(is.na(`Compressed Name`))

missingIndicators

# 
#missingIndicators %>% fwrite("../data/05_abbrevationmaker.csv",sep=";")

#rbind(compressedNames,addthis) %>% fwrite("../data/indicator_compressed_names.csv")
```



### plot function
```{r}

#label_modifier    = short_name_label_modifier,
## .overwrite = TRUE, # nolint: commented_code_linter.
.width  = 15
.height = 15

component_common <- "Dimension"
component_names <- paste(component_common, 1:7)
xy_expand <- function(xy, expand) xy * (1 + expand * expand_multiplier)
# Does not work
local_wdi_series<-local_wdi_series[grepl(":",local_wdi_series$Topic),]

local_wdi_series <- wdi_series_split_topic(local_wdi_series)

wdi_series_used_important <- local_wdi_series[`Indicator Code` %in% var_names]
var_names<-var_names[var_names%in% wdi_series_used_important$`Indicator Code`]
require(igraph)
stopifnot(length(var_names) == nrow(wdi_series_used_important))
dcor_var_iso_cons_subset<-dcor_var_iso_cons[rownames(dcor_var_iso_cons)%in% var_names,]
dcor_var_iso_cons<-dcor_var_iso_cons[rownames(dcor_var_iso_cons)%in% var_names,]



local_wdi_series<-local_wdi_series[local_wdi_series$`Indicator Code`%in% var_names,]
require(tidygraph)
require(ggraph)

dcor_subset<-dcor_var_iso_cons_subset
## local_wdi_series$`Indicator Code`
edges_level_01 <- local_wdi_series %>% select(`Indicator Code`, Topic1) %>%
  unique %>% rename( from = Topic1, to = `Indicator Code`)
edges_level_12 <- local_wdi_series %>% select(Topic1, Topic2)           %>%
  unique %>% rename( from = Topic2, to = Topic1          )
edges_level_23 <- local_wdi_series %>% select(Topic2, Topic3)           %>%
  unique %>% rename( from = Topic3, to = Topic2          )
edges_level_34 <- local_wdi_series %>% select(Topic3, Topic4)           %>%
  unique %>% rename( from = Topic4, to = Topic3          )
edges_level_45 <- local_wdi_series %>% select(Topic4, Topic5)           %>%
  unique %>% rename( from = Topic5, to = Topic4          )
## create the actual igraph object

rbind(edges_level_01, edges_level_12, edges_level_23, edges_level_34,
      edges_level_45) %>%
unique %>%
filter(to != from) ->listi
hier_graph <-listi %>% data.frame() %>%  { .[rev(seq_len(ncol(.)))] } %>% graph_from_data_frame
  ## graph_from_data_frame does not care about column names,
  ## only about positions
  # %>%
  #graph_from_data_frame

## additional data:

graph_idxs <- match(var_names, V(hier_graph)$name)

hier_graph <- set_vertex_attr(hier_graph, "which_max_iso", graph_idxs,
                              dcor_subset %>% apply(1, which.max))
hier_graph <- set_vertex_attr(hier_graph, "dcor_max", graph_idxs,
                              dcor_subset %>% apply(1, max))
hier_graph <- set_vertex_attr(hier_graph, "dcor1", graph_idxs,
                              dcor_subset[, 1])
hier_graph <- set_vertex_attr(hier_graph, "dcor2", graph_idxs,
                              dcor_subset[, 2])
hier_graph <- set_vertex_attr(hier_graph, "dcor3", graph_idxs,
                              dcor_subset[, 3])
hier_graph <- set_vertex_attr(hier_graph, "dcor4", graph_idxs,
                              dcor_subset[, 4])
hier_graph <- set_vertex_attr(hier_graph, "dcor5", graph_idxs,
                              dcor_subset[, 5])
hier_graph <- set_vertex_attr(hier_graph, "dcor6", graph_idxs,
                              dcor_subset[, 6])
hier_graph <- set_vertex_attr(hier_graph, "dcor7", graph_idxs,
                              dcor_subset[, 7])
# 

hier_graph <-
  set_vertex_attr(hier_graph, "Indicator.Code", graph_idxs,
                  local_wdi_series$`Indicator Code`[
                               match(var_names, local_wdi_series$`Indicator Code`)])



# thisPeople<-fread("../data/indicator_compressed_names.csv")
# wdi_names_prepared<-wdi_series_used_important[,c("Indicator Code","Indicator Name")]
# 
# 
# merge(wdi_names_prepared,thisPeople,by="Indicator Code",all.x = T)->wdi_names_prepared2

# wdi_names_prepared2 %>% select(`Indicator Code`,`Indicator Name`,`Compressed Name`)%>%  filter(is.na(`Compressed Name`)) %>% fwrite("../output/preparethenameschatgpt.csv")
# 
# wdi_series_used[,c("Indicator Code","Compressed Name")] #%>% fwrite("../data/indicator_compressed_names.csv")
# # # chatgpt shortened the names
# wdi_series_used
# wdi_series_used2<-fread("../data/indicator_compressed_names.csv")
# shortnames<-fread("../data/05_abbrevationmaker.csv",sep=";")
# rbind(wdi_series_used2[!(wdi_series_used2$`Indicator Code`%in% shortnames$`Indicator Code`)],shortnames) %>%  fwrite("../output/05_indicator_compressed_names.csv")

wdi_series_used<-fread("../data/indicator_compressed_names.csv")

wdi_var_hierarchy_graph <-hier_graph
axis_cons <- lapply(1:7, inner_connections_by_axis,
                       dcor_var_iso_cons       = dcor_var_iso_cons,
                       wdi_var_hierarchy_graph = wdi_var_hierarchy_graph,
                       var_names               = var_names)




expand_multiplier = 1/2
text_size         = 3
label_modifier    = manual_label_modifier
#label_modifier    = short_name_label_modifier
## .overwrite = TRUE, # nolint: commented_code_linter.
.width  = 50
.height = 50


wdi_var_hierarchy_graph %>%
ggraph(layout = "dendrogram", circular = TRUE) +
geom_node_point(aes(filter = leaf,
                    x = xy_expand(x, dcor1),
                    y = xy_expand(y, dcor1)),
                colour = col_pal[1]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor1),
                 y = xy_expand(y, dcor1)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[1]) +
geom_node_point(aes(filter = leaf,
                   x = xy_expand(x, dcor2),
                    y = xy_expand(y, dcor2)),
                colour = col_pal[2]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor2),
                 y = xy_expand(y, dcor2)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[2]) +
geom_node_point(aes(filter = leaf,
                    x = xy_expand(x, dcor3),
                    y = xy_expand(y, dcor3)),
                colour = col_pal[3]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor3),
                 y = xy_expand(y, dcor3)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[3]) +
geom_node_point(aes(filter = leaf,
                    x = xy_expand(x, dcor4),
                    y = xy_expand(y, dcor4)),
                colour = col_pal[4]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor4),
                 y = xy_expand(y, dcor4)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[4]) +
geom_node_point(aes(filter = leaf,
                    x = xy_expand(x, dcor5),
                    y = xy_expand(y, dcor5)),
                colour = col_pal[5]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor5),
                 y = xy_expand(y, dcor5)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[5]) +
  geom_node_point(aes(filter = leaf,
                    x = xy_expand(x, dcor5),
                    y = xy_expand(y, dcor5)),
                colour = col_pal[5]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor5),
                 y = xy_expand(y, dcor5)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[5]) +
  geom_node_point(aes(filter = leaf,
                    x = xy_expand(x, dcor5),
                    y = xy_expand(y, dcor5)),
                colour = col_pal[6]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor5),
                 y = xy_expand(y, dcor5)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[6]) +
      geom_node_point(aes(filter = leaf,
                    x = xy_expand(x, dcor5),
                    y = xy_expand(y, dcor5)),
                colour = col_pal[7]) +
geom_polygon(aes(sort = name,
                 filter = name %in% var_names,
                 x = xy_expand(x, dcor5),
                 y = xy_expand(y, dcor5)),
             stat = stat_filter_sort,
             group = 1,
             fill = NA,
             colour = col_pal[7]) +
geom_node_point(aes(filter = leaf, color = factor(which_max_iso)),
                size = 0)+
geom_node_text(aes(filter = leaf,
                   label  = label_modifier(as.character(Indicator.Code)),
                   color  = factor(paste(component_common, which_max_iso)),
                   x      = x * (1 + 1 * expand_multiplier),
                   y      = y * (1 + 1 * expand_multiplier),
                   hjust  = if_else(x > 0, 0, 1),
                   angle  = atan(y / x) / 2 / pi * 360),
               show.legend = FALSE,
               size        = text_size) +
geom_conn_bundle(data = get_con(axis_cons[[1]][1, ], axis_cons[[1]][2, ]),
                 edge_colour = col_pal[1], edge_alpha = 0.3) +
geom_conn_bundle(data = get_con(axis_cons[[2]][1, ], axis_cons[[2]][2, ]),
                 edge_colour = col_pal[2], edge_alpha = 0.3) +
geom_conn_bundle(data = get_con(axis_cons[[3]][1, ], axis_cons[[3]][2, ]),
                 edge_colour = col_pal[3], edge_alpha = 0.3) +
(
  if (ncol(axis_cons[[4]]) > 0)
    geom_conn_bundle(data = get_con(axis_cons[[4]][1, ],
                                    axis_cons[[4]][2, ]),
                     edge_colour = col_pal[4], edge_alpha = 0.3)
  else
    list()
) +
geom_conn_bundle(data = get_con(axis_cons[[5]][1, ], axis_cons[[5]][2, ]),
                 edge_colour = col_pal[5], edge_alpha = 0.3) +
geom_edge_diagonal() +
# geom_node_label(aes(filter = !leaf,
#                     fontface = if_else(name %in% level1, "bold", "plain"),
#                     label = name, angle = atan(y / x) / 2 / pi * 360),
#                 size = 2) +
scale_color_manual(limits = component_names,
                   labels = component_names,
                   values = col_pal) +
guides(colour = guide_legend(title = component_common,
                             override.aes = list(size = 5))) +
theme_void() +
theme(legend.title = element_blank(),
      legend.position = c(0.9, 0.1),
      legend.justification = c(0.9, 0.1)) +
expand_limits(x = c(-2.7, 2.7), y = c(-2.7, 2.7))

ggsave("../figures/07_230808_circle_testshaplez.pdf",device = "pdf",height = 12,width = 12)

```



# Result 6

Conrol THE outliers
```{r}
#subsetwdi
num_levels <- 4
# Use cut() to create levels of quantiles
dimred2021_knn250_2$pscore_level<-as.factor(ntile(dimred2021_knn250_2$pscore_mean, num_levels))
dimred2021_knn250_2 %>% group_by(pscore_level) %>% summarize(Min=min(pscore_mean),Max=max(pscore_mean)) %>% mutate(p_score_level_label=paste0(Min," to ",Max)) %>% 
  left_join(dimred2021_knn250_2, . ,by="pscore_level")-> dimred2021_knn250_3
dimred2021_knn250_3 %>%filter(year%in% 2020) %>%   select(`Country Code`,`Short Name`,TX.VAL.MRCH.R3.ZS,Region.x) %>%  arrange(desc(TX.VAL.MRCH.R3.ZS))
dimred2021_knn250_3 %>% filter(year%in% 2020) %>%  select(`Country Code`,`Short Name`,SL.TLF.CACT.FE.ZS,Region.x) %>%  arrange(desc(SL.TLF.CACT.FE.ZS))

```


```{r}
dimred2021_knn250_3 %>% 
  # mutate(pscore_cat2 = fct_relevel(pscore_cat, 
  #           "", "<0", "0 to <5","5 to <10","10 to <20","20 to <30","30 to <40","40 to <50", ">=50" )) %>% 
  mutate(p_score_level_label = fct_relevel(p_score_level_label,
            "-6.4 to 5.5" , "5.5 to 8.6" , "8.6 to 18.1" ,"18.1 to 97.1"  )) %>%
  mutate(p_score_level_label2=case_when(p_score_level_label%in% "18.1 to 97.1" & RegionSoviet %in% "Post-Soviet"~"High Soviets",
                                            p_score_level_label%in% "18.1 to 97.1" & RegionSoviet %in% "Latin America & Caribbean"~"High Latin",
                                               p_score_level_label%in% "18.1 to 97.1" ~"High other",
                                              TRUE~p_score_level_label ))    %>%                                                                                     
  # mutate(p_score_level_label = fct_relevel(p_score_level_label, 
  #           "-6.4 to 4.4","4.4 to 7" ,  "7 to 11.6" ,"11.6 to 20.1" ,"20.1 to 97.1" )) %>% 

  filter(!is.na(pscore_level))->dimred2021_knn250_3 
dimred2021_knn250_3$p_score_level_label2 %>%  table

levels(as.factor(dimred2021_knn250_3$p_score_level_label))
Indicator<-"EN.ATM.CO2E.KD.GD"
year_start<-"2015"
year_end<-"2021"

for(i in 1:nrow(subsetwdi)){
  Indicator<-subsetwdi[i,"Series Code"] %>%  unlist() %>%  as.character()
  
  a<-get_wdi_trajectory(data=dimred2021_knn250_3,Indicator,groupedby ="p_score_level_label2" ,CI = T)
  a %>% ggsave(file = paste0("../test/trajectories/split_regions/",str_replace_all(Indicator,"\\.","_"),".png"),device = "png")
}

dimred2021_knn250_5 %>%  filter(p_score_level_label2%in% "High other") %>% select(`Country Code`,RegionSoviet,`Country Name`) %>%  unique-> asd
```


```{r}
a<-get_wdi_trajectory(data=dimred2021_knn250_3,"iso1","p_score_level_label2",CI = T)
b<-get_wdi_trajectory(data=dimred2021_knn250_3,"iso2","p_score_level_label2",CI = T)
c<-get_wdi_trajectory(data=dimred2021_knn250_3,"iso3","p_score_level_label2",CI = T)
d<-get_wdi_trajectory(data=dimred2021_knn250_3,"iso4","p_score_level_label2",CI = T)

ggarrange(a+theme(legend.position = "asd")+xlab("")+ggtitle("iso1"),b+theme(legend.position = "asd")+xlab("")+ggtitle("iso2"),c+theme(legend.position = "asd")+xlab("")+ggtitle("iso3"),d+theme(legend.position = "asd")+xlab("")+ggtitle("iso4")) %>%  ggsave(file="../figures/07_230810_isoExcess.png",device = "png")
```
#### High former Soviet statets

add indicators that are not in the dimred
```{r}
iso2allcors<-fread("../results/corrTable/05_CorrTable_iso2_quantiles_shapleys_2019to2021_all.csv")
iso2allcors$iso2_all %>%  summary()
wdi_data<-fread("../data/05_WDI_data.csv")


add_col<-c(colnames(wdi_data)[!(colnames(wdi_data)%in% colnames(dimred2021_knn250_3))])

wdi_data %>% mutate(identifier=paste0(`Country Code`,year)) %>%select(any_of(c("Country Code","year","identifier",add_col)))-> wdi_data2
dimred2021_knn250_3 %>% mutate(identifier=paste0(`Country Code`,year)) -> dimred2021_knn250_4

left_join(dimred2021_knn250_4,wdi_data2 %>% select(any_of(c("identifier",add_col))),"identifier") %>%  filter() -> dimred2021_knn250_5

iso2allcors %>% filter(N_75to100_75to100>mean(N_75to100_75to100)) %>%  arrange(desc(iso2_75to100_75to100)) %>% slice(1:100)-> newsubset
newsubset->newsubset
for(i in 1:nrow(newsubset)){
  Indicator<-newsubset[i,"Series Code"] %>%  unlist() %>%  as.character()
  a<-get_wdi_trajectory(data=dimred2021_knn250_5,Indicator,groupedby ="p_score_level_label2" ,CI = T)
  
  a %>% ggsave(file = paste0("../test/trajectories/ISO2_forsoviets/",str_replace_all(Indicator,"\\.","_"),".png"),device = "png")
}
  a<-get_wdi_trajectory(data=dimred2021_knn250_5,"SP.POP.DPND.OL",groupedby ="p_score_level_label2" ,CI = T)
a
dimred2021_knn250_5 %>% filter(year%in% 2020)%>%  arrange(desc(SP.POP.DPND.OL)) %>% select(RegionSoviet,iso1,iso2,iso3,iso4)
```















# SUPPLEMENT

# 3 Figure nonlinear isomap types

```{r}
wdi_data_cons_df<-wdi_data_cons_df

featureList<-c("NY.GDP.PCAP.CD","Country Code","P_mort_2020","P_mort_2021","iso1","iso2","iso3","iso4","iso5","iso6","iso7","iso8","iso9","iso10","year","pscore_mean","RegionSoviet","Income Group")
```


## Plot gdp iso 1
```{r}
subset2020<-dimred2021_knn250_2 %>% filter(year%in% 2020) %>% select(any_of(featureList)) %>% rename(ExcessMort=P_mort_2020) %>%  mutate(P_mort_2020=NULL)
subset2021<-dimred2021_knn250_2 %>% filter(year%in% 2021) %>% select(any_of(featureList)) %>% rename(ExcessMort=P_mort_2021) %>% mutate(P_mort_2021=NULL)

bothyears<-rbind(subset2020,subset2021,use.names=FALSE)


bothyears %>% filter(year%in% 2021) %>% mutate(Income_group=`Income Group`) %>% filter(!is.na(Income_group))%>% 
  mutate(Income_group = factor(Income_group, levels=c("High income",
                                      "Upper middle income",
                                      "Lower middle income",
                                      "Low income"
                                      )),`GDP per Capita (log)`=log(NY.GDP.PCAP.CD),
         `Isomap dimension 1`=iso1
         ) %>% 
  
   pivot_longer(names_to = "var",
               c("Isomap dimension 1","GDP per Capita (log)"),
               values_drop_na = TRUE) %>% 
  ggplot(aes(x=(value),y=ExcessMort,color=Income_group))+geom_point()+ theme_bw()+xlab("")+ylab("Excess Mortality") +scale_color_viridis(discrete = T)+facet_wrap(~var,scales = "free_x")+theme(legend.position = "bottom") 
#ggsave("../figures/06_figures_GDP_Iso1.png",device = "png", width = 7, height = 5, dpi = 300, units = "in")
```
## Plot iso 2

### check all the plots
```{r}

wdi_dimred_iso2 %>% filter(!is.na(iso2_all))%>%  mutate(maxdisthighlow=abs(iso2_0to25-iso2_50to75),
                                                        adjustvalue=iso2_75to100/maxdisthighlow) %>% arrange(desc(maxdisthighlow))-> subsetiso2

# for(i in 1:676){
# nameVar<-subsetiso2[i,"Series Code"] %>%  unlist
# applySD<-!is.na(dimred2021_knn250_2[,..nameVar]) %>%  unlist
# 
# wool<-dimred2021_knn250_2[applySD[,1]] %>% filter(year%in% 2010:2021) %>% filter(!is.na(iso2)) %>% 
# ggplot(aes_string(x="iso2",y=paste(nameVar)))+geom_point()
# wool %>% ggsave(file = paste0("../test/run",i,".png",sep=""),device = "png")
# }

```
#### candidates
```{r}
iso2_candidates<-c(5,15,21,26,27,32,49,57,66,68,70,71,80,97,102,119,123,180,207,377,383,407,439,583)
nameVar_iso2_candidates<-subsetiso2[iso2_candidates,"Series Code"] 

for(i in iso2_candidates){
nameVar<-subsetiso2[i,"Series Code"] %>%  unlist
applySD<-!is.na(dimred2021_knn250_2[,..nameVar]) %>%  unlist

wool<-dimred2021_knn250_2[applySD[,1]] %>% filter(year%in% 2010:2021) %>% filter(!is.na(iso2))


wool %>%  mutate(categorie=case_when(scale(iso2)>1.5~"Iso2 peak  high",
                                                 scale(iso2)<  -0.5~"Iso2 peak  Low",
                                                 (scale(iso2)> (-0.5) &scale(iso2)<1.5)~"normal",
                                                  TRUE~"a"
                                                  ))%>% 
  filter(!is.na(get(nameVar)),!is.na(categorie)) -> yo


yo$categorie<-factor(as.factor(yo$categorie),levels=c("Iso2 peak  Low","normal","Iso2 peak  high","a"))
yo2<-yo[,colnames(yo)[colnames(yo)%in% c(paste(nameVar),"categorie")],with=F]
colnames(yo2)[1]<-"value"
series_label<-wdi_explained[wdi_explained$`Series Code` %in% nameVar,"Indicator Name"]

wool2<-ggstatsplot::ggbetweenstats(
  plot.type = "boxviolin",
                            p.adjust.method = "BH",
        point.args = list(alpha = 0),

      data = yo2,
      x = categorie,
      y = value ) + 
      # Add labels and title
      labs(x="",
           y = nameVar,
           title = series_label
      )+theme(
        # This is the new default font in the plot
        text = element_text(family = "Roboto", size = 10, color = "black"),
        plot.title = element_text(
          family = "Lobster Two", 
          size = 14,
          face = "bold",
          color = "#2a475e"
        ),
        # # Statistical annotations below the main title
        plot.subtitle = element_text(
          family = "Roboto",
          face = "bold",
          color="#1b2838"
        ),
        plot.title.position = "plot", # slightly different from default
        axis.text = element_text(size = 10, color = "black"),
        axis.title = element_text(size = 12)
      )

wool2 %>% ggsave(file = paste0("../figures/06_iso2plots/",i,".png",sep=""),device = "png")
}

```

### plots
```{r}
featureList<-c("SH.STA.OWGH.ME.ZS","VC.IHR.PSRC.MA.P5","TM.VAL.MRCH.R3.ZS","TX.VAL.MRCH.R2.ZS","SP.DYN.CDRT.IN","SP.POP.2024.FE.5Y","Country Code","P_mort_2020","P_mort_2021","iso1","iso2","iso3","iso4","iso5","iso6","iso7","iso8","iso9","iso10","year","pscore_mean","RegionSoviet","Income Group")


#
# 
# ggsave("../figures/06_figuressa_GDP_Iso1.png",device = "png", width = 7, height = 5, dpi = 300, units = "in")
```
```{r}
dimred2021_knn250_2 %>% filter(year%in% 2019:2021) %>% mutate(Income_group=`Income Group`) %>% filter(!is.na(Income_group)) %>% 
  ggplot(aes(x=(iso2),y=(SP.DYN.CDRT.IN)))+geom_point()+geom_smooth(method = "lm")+ theme_bw()+xlab("Dimension 2")+ylab("Death rate, crude (per 1,000 people)") +scale_color_viridis(discrete = T) 


```
```{r}
dimred2021_knn250_2 %>% filter(year%in% 2019:2021) %>% mutate(Income_group=`Income Group`) %>% filter(!is.na(Income_group)) %>% 
  ggplot(aes(x=(iso2),y=TM.VAL.MRCH.R3.ZS))+geom_point()+geom_smooth(method = "lm")+ theme_bw()+xlab("Dimension 2")+ylab("Merchandise imports from low- and middle-income economies in Latin America & the Caribbean (% of total merchandise imports)") +scale_color_viridis(discrete = T) 
dimred2021_knn250_2 %>% filter(year%in% 2019:2021) %>% mutate(Income_group=`Income Group`) %>% filter(!is.na(Income_group)) %>% 
  ggplot(aes(x=(iso2),y=TX.VAL.MRCH.R2.ZS))+geom_point()+geom_smooth(method = "lm")+ theme_bw()+xlab("Dimension 2")+ylab("Merchandise exports to low- and middle-income economies in Europe & Central Asia (% of total merchandise exports)") +scale_color_viridis(discrete = T) 
```

## Plot iso 1

### check all the plots
```{r}
wdi_dimred_iso1<-fread("../results/corrTable/04_CorrTable_iso1_quantiles_all.csv")
wdi_dimred_iso1 %>% filter(!is.na(iso1_all))%>%  mutate(maxdisthighlow=abs(iso1_0to25-iso1_50to75),
                                                        adjustvalue=iso1_75to100/maxdisthighlow) %>% arrange(desc(maxdisthighlow))-> subsetiso1
```


```{r}
for(i in 1:100){
nameVar<-subsetiso1[i,"Series Code"] %>%  unlist
applySD<-!is.na(dimred2021_knn250_2[,..nameVar]) %>%  unlist

wool<-dimred2021_knn250_2[applySD[,1]] %>% filter(year%in% 2010:2021) %>% filter(!is.na(iso2)) %>%
ggplot(aes_string(x="iso1",y=paste(nameVar)))+geom_point()
wool %>% ggsave(file = paste0("../test/run",i,".png",sep=""),device = "png")
}

```
#### candidates
```{r}
iso1_candidates<-c(24,3,34,58)
nameVar_iso1_candidates<-subsetiso1[iso1_candidates,"Series Code"] 

indicator_a<-nameVar_iso1_candidates$`Series Code`[1]

a<-dimred2021_knn250_2 %>% filter(year%in% 2000:2021) %>% 
  ggplot(aes_string(x="iso1",y=indicator_a))+geom_point(color="firebrick4")+ggtitle(paste(wdi_explained[wdi_explained$`Series Code`%in% indicator_a,"Indicator Name"]))+theme_bw()

indicator_b<-nameVar_iso1_candidates$`Series Code`[2]

b<-dimred2021_knn250_2 %>% filter(year%in% 2000:2021) %>% 
  ggplot(aes_string(x="iso1",y=indicator_b))+geom_point(color="blue4")+ggtitle(paste(wdi_explained[wdi_explained$`Series Code`%in% indicator_b,"Indicator Name"]))+theme_bw()
ggarrange(a,b,ncol = 1) %>% ggsave(file="../figures/06_figures_iso1_nonlinear.png",device = "png", width = 7, height = 5, dpi = 300, units = "in")
```

## Plot iso 4

### check all the plots
```{r}
wdi_dimred_iso4<-fread("../results/corrTable/04_CorrTable_iso4_quantiles_all.csv")
wdi_dimred_iso4 %>% filter(!is.na(iso4_all_all))%>%  mutate(maxdisthighlow=abs(iso4_0to25-iso4_75to100),
                                                        adjustvalue=iso4_50to75/maxdisthighlow) %>% arrange(desc(maxdisthighlow))-> subsetiso4
```


```{r}
for(i in 1:100){
nameVar<-subsetiso4[i,"Series Code"] %>%  unlist
applySD<-!is.na(dimred2021_knn250_2[,..nameVar]) %>%  unlist

wool<-dimred2021_knn250_2[applySD[,1]] %>% filter(year%in% 2010:2021) %>% filter(!is.na(iso4)) %>%
ggplot(aes_string(x="iso4",y=paste(nameVar)))+geom_point()
wool %>% ggsave(file = paste0("../test/run",i,".png",sep=""),device = "png")
}

```
#### candidates
```{r}
iso1_candidates<-c(24,3,34,58)
nameVar_iso1_candidates<-subsetiso1[iso1_candidates,"Series Code"] 

indicator_a<-nameVar_iso1_candidates$`Series Code`[1]

dimred2021_knn250_2 %>% filter(year%in% 2000:2021) %>% 
  ggplot(aes_string(x="iso1",y=indicator_a))+geom_point()+ggtitle(paste(wdi_explained[wdi_explained$`Series Code`%in% indicator_a,"Indicator Name"]))
```


# 4 Shapley values
